tests: add technology-specific feature files
Separates the test feature files into different subfolders depending on
the workflow engine (cwl, yadage, snakemake) used by the workflow. This
is needed to look for log messages that are relevant to the specific
workflow engine, and because workflows with different engines create
different files in different folders of the workspace.
Tests that are valid for all the workflow engines are kept at the
highest level in the `tests` directory.